Extract the synchronizer methods from the participant admin connection
and rename everything to synchronzier

[ci]
