+++ apps/sv/src/main/scala/org/lfdecentralizedtrust/splice/sv/automation/singlesv/SyncConnectionStalenessCheck.scala
@@ -13,5 +13,5 @@ trait SyncConnectionStalenessCheck {
   def participantAdminConnection: ParticipantAdminConnection
 
   def isNotConnectedToSync()(implicit tc: TraceContext, ec: ExecutionContext): Future[Boolean] =
-    participantAdminConnection.listConnectedDomains().map(_.isEmpty)
+    participantAdminConnection.listConnectedSynchronizers().map(_.isEmpty)
 }
