+++ apps/sv/src/main/scala/org/lfdecentralizedtrust/splice/sv/onboarding/joining/JoiningNodeInitializer.scala
@@ -165,7 +165,7 @@ class JoiningNodeInitializer(
         // If the url is unset, we skip this step. This is fine if the node has already initialized its
         // own sequencer.
         domainConfigO.traverse_(
-          participantAdminConnection.ensureDomainRegisteredNoHandshake(
+          participantAdminConnection.ensureSynchronizerRegisteredNoHandshake(
             _,
             RetryFor.WaitingOnInitDependency,
           )
@@ -523,7 +523,7 @@ class JoiningNodeInitializer(
             dsoPartyToParticipantMapping.nonEmpty || activeDsoPartyToParticipantProposals.isEmpty
           ) {
             logger.info("Reconnecting all domains.")
-            participantAdminConnection.reconnectAllDomains()
+            participantAdminConnection.reconnectAllSynchronizers()
           } else {
             Future.unit
           }
